LPS-123793 Rename Renderer component to Page and consumes data directly from useFormState and useConfig
